upload_api/requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Cython==0.29.32
elastic-transport==8.4.0
elasticsearch7==7.17.6
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
layoutparser==0.3.4
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pdf2image==1.16.0
pdfminer.six==20220524
pdfplumber==0.7.4
Pillow==9.2.0
portalocker==2.5.1
pycparser==2.21
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
six==1.16.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Wand==0.6.10
Werkzeug==2.2.2
zipp==3.8.1
